+++ torcherino/src/main/java/torcherino/ModContent.java
@@ -30,7 +30,7 @@ import torcherino.block.tile.CustomTileEntityType;
 import torcherino.block.tile.TorcherinoTileEntity;
 
 @Mod.EventBusSubscriber(modid = Torcherino.MOD_ID, bus = Mod.EventBusSubscriber.Bus.MOD)
-public class ModContent
+public final class ModContent
 {
     private static final DeferredRegister<Block> BLOCKS = DeferredRegister.create(ForgeRegistries.BLOCKS, Torcherino.MOD_ID);
     private static final DeferredRegister<Item> ITEMS = DeferredRegister.create(ForgeRegistries.ITEMS, Torcherino.MOD_ID);