+++ torcherino/src/main/java/torcherino/block/TorcherinoBlock.java
@@ -22,13 +22,13 @@ import net.minecraft.world.IBlockReader;
 import net.minecraft.world.World;
 import net.minecraft.world.server.ServerWorld;
 import org.apache.commons.lang3.StringUtils;
-import org.jetbrains.annotations.Nullable;
 import torcherino.Torcherino;
 import torcherino.api.TierSupplier;
 import torcherino.block.tile.TorcherinoTileEntity;
 import torcherino.config.Config;
 import torcherino.network.Networker;
 
+import javax.annotation.Nullable;
 import java.util.Random;
 
 import static net.minecraft.state.properties.BlockStateProperties.POWERED;