+++ torcherino/src/main/java/torcherino/network/Networker.java
@@ -12,7 +12,7 @@ import torcherino.Torcherino;
 import torcherino.api.TorcherinoAPI;
 import torcherino.block.tile.TorcherinoTileEntity;
 
-public class Networker
+public final class Networker
 {
     public static final Networker INSTANCE = new Networker();
 
@@ -20,7 +20,7 @@ public class Networker
 
     public void initialise()
     {
-        String version = "2";
+        final String version = "2";
         torcherinoChannel = NetworkRegistry.newSimpleChannel(Torcherino.getRl("channel"), () -> version, version::equals, version::equals);
         torcherinoChannel.registerMessage(0, ValueUpdateMessage.class, ValueUpdateMessage::encode, ValueUpdateMessage::decode, ValueUpdateMessage::handle);
         torcherinoChannel.registerMessage(1, OpenScreenMessage.class, OpenScreenMessage::encode, OpenScreenMessage::decode, OpenScreenMessage::handle);
@@ -30,14 +30,14 @@ public class Networker
     public void openScreenServer(final World world, final ServerPlayerEntity player, final BlockPos pos)
     {
         if (world.isRemote) { return; }
-        TileEntity tile = world.getTileEntity(pos);
+        final TileEntity tile = world.getTileEntity(pos);
         if (tile instanceof TorcherinoTileEntity)
         { torcherinoChannel.sendTo(((TorcherinoTileEntity) tile).createOpenMessage(), player.connection.netManager, NetworkDirection.PLAY_TO_CLIENT); }
     }
 
     public void sendServerTiers(final ServerPlayerEntity player)
     {
-        S2CTierSyncMessage message = new S2CTierSyncMessage(TorcherinoAPI.INSTANCE.getTiers());
+        final S2CTierSyncMessage message = new S2CTierSyncMessage(TorcherinoAPI.INSTANCE.getTiers());
         torcherinoChannel.sendTo(message, player.connection.netManager, NetworkDirection.PLAY_TO_CLIENT);
     }